requirements.txt
pillow==7.2.0
pytesseract==0.3.4
Flask==1.1.2

